MySQLTable: Fix escaping bug in .insert() and .update()
Fixes a bug where if the first argument passed to .insert() or .update() was
an object and one of the values was a string with a question mark, the
question mark would be used as an escaping placeholder.

This fix required the signature to .update() to be changed. Now the `data`
argument cannot be a string.