lib: Fix foreign key bug introduced in d43235a56772115ac1b945cc670f7ffd8f76c447
That commit used a boolean which caused an "ADD FOREIGN KEY" statement to be created
for each foreign key in a table if at least one of them needed to be added back after
dropping it to prevent conflicts. This commit fixes the logic so that only the
specific keys that were dropped get added back.