PoolPlus: Add debugging to #sync() and fix incorrect documentation
The documentation suggested that if the sync failed, all changes to the database tables would be reverted.
This was not true because even though the code used a MySQL transaction, the types of SQL statements used
to migrate the database schema cause an implicit commit, so using a transaction does absolutely nothing.
The code has now been changed to use a regular connection instead of a transaction.